14222 BI configuration: Changed element order of "restrict severity to at worst"
This werk only introduces a visual fix, so no functional changes.
The order of the dropdown choice elements did not reflect the severity of the states.
The correct severity order for the BI is OK->WARN->UNKNOWN->CRIT, which differs
from the order of the monitoring states OK->WARN->CRIT->UNKNOWN.

Change-Id: Icbeef8b8a5fc3391698814b0d7c3f40472461aee