unbreak nagios core config generation
In the process of preparing a bulk migration of the legacy
check plugins we renamed a couple of them, to remove some
... irregularities of snmp sections.

This revealed that the nagios core config generation actually
relied on the fact that every check plugin must be defined in
a file with the same name.

We now track where a plugin has been discovered, which is
a) more consistent with the modern approach, and b) more
robust.

CMK-15790

Change-Id: I6492dfde305052d56befc5e020dba3e743e1e707