python3-modules: Write pip install output to a log
To be able what is going on with the pip command, we write the output to
a log file. This way we can investigate what is happening while the
command is being executed or later on, even when the command succeeded.

It seems we have no other chance to get the output of from the pip
execution while it is running.

CMK-15923

Change-Id: Ie1318cf702489db394c86023a5bdf5433e93c17d