extension package tests: move update check
We move the update check into the pytest scope.
Putting a little more effort in the update check
allows us to move from asserting

"Ensure we are testing the n most downloaded MKPs"

to

"Ensure the n MKPs we are testing are in the m
most downloaded ones",

which gives us a little more freedom in how
often we have to update this test when download
counts change.

Change-Id: I52489c8ed67ebe87757245a954561346ab84068e