skip old legacy check info elements
Making the LegacyCheckDefinition a dataclass
allows for a runtime instance check.

This way we can filter out all plugins that
are not in line with what ever this
no-longer-stable, deprecated API chooses to
accept, avoiding havoc during update config.

Not loading local legacy checks at all is
not an option, because there might be
argument thingies in the files.

Loading local legacy checks but then throwing
away the new contents of the check_info is
not very good, because it will prevent local
"hot patches" of legacy plugins.

Change-Id: I6ff6a50cff34f47adc453e9b4e186fb4d5606c1f