Let install-rust-cargo install multiple version
This updates the default channel to 1.75 and keeps 1.72 around as
additional version.

We keep one default version but are now able to install multiple
additional versions in case these are required.

Versions 1.75 and 1.76 of the rust libs showed problems when their
libs have been stripped. We are not stripping them.

Change-Id: Ica19803c3d8b5e6b3d4215a8e081769f2b0abe36