16365 FIX baked_agents: Fix version displayed on agent files
Prior to this Werk, the running version of Checkmk was used to display
the version of the agents. This caused that after updating Checkmk to a
new version, the wrong version of the agents was displayed until they
were baked again.

To fix this bug, the agent version is now displayed instead of the
Checkmk version.

SUP-14846

Change-Id: Ib007875f9675cb102cde3f76ac4c669f0fc0a340