16466 introduce possibility to update discovered service labels
Prior to this werk, it was not possible to update only the
discovered labels of a service. Updating those were tied
to updating the identified parameters of the service. This
werk introduces the option to update the discovered service
labels only automatically as well as manually. The following
changes were made to introduce this option as well for better
visibility:

* Service discovery page: changed services due to a change in
their labels are now listed under 'Changed services'. Service
labels can be manually updated with dedicated actions targeting
all changed services or a specific service
* Check_MK Discovery service: the service now accounts for
changed services due to label changes. The state of the discovery
service for such cases can also be configured
* Bulk discovery & Periodic service discovery: now have the option
to update labels of changed services

Change-Id: I2e7fd20d80fe15307d19473810d18718d07dd465