Move spec generating code to dedicated module
The end goal is to extract the whole spec computation out of the GUI
application. The intermediate target is to isolate all the code we need
to generate the spec in a dedicated module which is also only imported
when needed.

This change moves the generic piece of the code to a new module where we
will move a few other pieces to in the folloing commits.

CMK-15981

Change-Id: I77dccc7419305049dc1e95486fb4d106435f5f4f