16391 FIX ldap_connection: When you create/clone a new ldap connection, the ID will now be unique
Previously, when creating a new LDAP connection via the UI, you could provide
an ID that was already in use by another user connection. This is no longer
the case with this werk. Now, when you clone a connection, a new unique ID
will be generated from the cloned connection. If you create a new connection,
the ID must be unique or you will get an error.

Change-Id: I12489a87ff51c36cbd9a1c79c046cedfd0073cd8