16754 FIX docker_container: skip on incomplete data for diskstat and memory
During the data parsing of a container, it is possible to encounter
incomplete metric sets, which previously caused a crash. Since the
data in these instances is simply unavailable, we now skip the
discovery or check cycle for such cases. This adjustment is applied
to docker container disktat and memory check plugins.

SUP-17452
CMK-16933
CMK-16934

Change-Id: Ifd1cb9734839d44177a3a5519f126646085dbe9d