16393 FIX notification_rule: allow custom plugin names when selecting cancel without a restart
We recently made a change that would allow you to create a notification rule,
via the REST-API using a custom plugin but also setting the option
"cancel_previous_notifications".  However, an apache restart was required
since we were verifying the custom plugin via the runtime generated
schema. This werk fixes this issue by removing this verification from the
schema and checking at the endpoint.

Change-Id: I61b6ce9449a9cffe096c86d13fcb13a129650259