add possibility to display up/down state for 'if'
On windows the information is send in the section
'winperf_if' may be extended with a pseudo counter
IF_STATUS_PSEUDO_COUNTER = "2002"

Previously any network interface had been assumed as
Connected.

If pseudo counter is absent, then we have got original
behavior but with `up` instead of `Connected` as
require by ticket.

CMK-15220

Change-Id: I5e9ce1c2bde8e7fa1eec31ee7e2da15eeefdad06