add possibility to display up/down state for 'if'
On windows the information is send in the section
'winperf_if' may be extended with a pseudo counter
IF_STATUS_PSEUDO_COUNTER = "2002"

Previously any network interface had been assumed as
Connected.

If pseudo counter is absent, then we have got original
behavior but with `up` instead of `Connected` as
require by ticket.

CMK-15220

Change-Id: I7207685035552988e03e76928031c6d843fa4e1b