16520 FIX Bi: creating rules should allow the old style host_labels or service_labels
Our host labels and service labels recently went through some
major changes which included updating the REST-API fields to
host_label_groups and service_label_groups.

However, for reasons of backward compatibility, we decided to keep the
legacy host_labels and service_labels as possible fields when
creating a bi rule via the REST-API. This werk implements this change,
but these legacy fields will be marked as deprecated and will not be
available in our next major release. It is recommended to use the
new format.

Change-Id: I9632eb1779abed85ff4ac2874afe7a8f6e38a151