Add dedicated command to compute REST API spec: cmk-compute-api-spec
Previously the REST API specification was computed and cached in every
apache process. This caused several issues:

a) After spawning a new apache, the spec needed to be recomputed for
   every process. This caused a delay in the first request hitting an
   apache process asking for it.
b) It was held in memory by every process consuming a few MB.
c) The invalidation of the cache and computation of new spec could not
   be triggered manually.

With this change the spec is now stored in the site and made available
to all apache processes from there.

With the dedicated command `cmk-compute-api-spec` the computation can
now be triggered in specific situations automatically or manually for
debugging.

CMK-15981

Change-Id: I7c7a04b15d7715024f4ef213689edbf0ad630795