update config: autochecks: separate pre- and actual action
We have to somehow deal with autocheck enrties that we can't
properly migrate.

We now have two actions for the rewriting of the autochecks:

1) A pre-update action, showing warnings to the user and
offering the chance to "abort and fix manually", but not
persisting any changes

2) An actual update action, swallowing all errors and simply
droping the corresponding services. In exchange we get an
autochecks file that is guaranteed to be readable.

CMK-16206

Change-Id: I920e5481869d2a94a4d04e1a0de57f0fe7605541