Background job for spec creation triggered by UI actions
We need to trigger an update of the spec to reflect changes made by previous
config steps in the API specification.

Since the computation takes some time, this must not be done
synchronously. We need to be able to track the outcome of the job and be
able to analyze the progress of the job through the UI. So it makes
sense to realize this as background job.

The job shall be triggered by every change to the configuration which
has an effect on the API specification. The first place we trigger this
job is the configuration of tags.

Change-Id: Ie595552cd79eb3c38abdfa594b91ac7c99ac9330