Fix "omd mv"
In 2.3 "omd mv" failed to update the files managed by the skel
mechanism with this error message:

```
Moving site lmtest3 to lmtest2...
Error patching template file '/omd/sites/lmtest2/./.bashrc': Failed to determine site version
(...)
```

The replacements of the "old site" could previously not be determined
correctly because the files of the site, including the version symlink,
were already moved to the new site name. With the version symlink being
moved, the ###EDITION### macro could not be computed anymore.

Change-Id: Ieabf35849ba9174ce699de75f7963c3791d00f02