16521 FIX bi_rule: schema update to match the api docs
The Open API schema previously did not reflect the response or the request
schema format that was required to create or show bi_rules. This werk
addresses this issue.

Previously, when creating or getting a bi-rule, via the REST-API, the
schema for host_label_groups or service_label_groups looked something
like this

```
    "host_label_groups": [
        [
            "and",
            [
                ["and", "mystery/switch:yes"],
                ["or", "mystery/switch:no"],
            ],
        ],
    ]
```

This did not match the schema documented in the Open API docs.
To fix this, we have now changed the format to the following

```
    "host_label_groups": [
        {
            "operator": "and",
            "label_group": [
                {"operator": "and", "label": "mystery/switch:yes"},
                {"operator": "or", "label": "mystery/switch:no"},
            ],
        },
    ]
```

This also aligns with other endpoints that use our new
host_label_groups or service_label_groups, for example the
rules endpoints.

As this is a breaking change, user scripts should be adjusted
accordingly.

Change-Id: I14657ad8bf37e71c69210b9d7e4f2dfaa2f4d94d