Execute spec endpoint tests with dummy spec
Previously these tests were meant to ensure the endpoints work and
whether the API spec can be computed. This lead to long execution times.
Since the spec is now generated separately, we can also execute the tests
with a dummy spec and can remove the slow marker.

For the computation of the spec in the site environment we have an
integration test in place. Of course, this is not ideal for the
development workflow. So we need to find a way to test the spec
computation in a subsequent step.

Cleanup `validate_spec` related warnings along the way.

CMK-15981

Change-Id: Id8a642fa4b5aaa55beff9a87df76a708e7b14308