15839 Connection state of interface on Windows is supported out-of-the-box
Previously, users were required to install an additional plugin to
determine the state of a network interface on Windows.

With the current release, this functionality is now integrated directly into the
Windows agent, eliminating the need for extra plugins.

For enhanced consistency across operating systems, network interfaces that are
operational are now marked as `up`, and those that are not available are marked
as `down`, aligning with the terminology used on Linux.

Change-Id: I5da43659ed0eca74653059a5d293ed38db2980da