16500 FIX service_discovery: allow discovery on fresh remote host
The werk 16465 addressed a problem that prevented fetching
information about the service discovery background job on a
remote site. However, this solution introduced a new limitation,
disallowing the execution of discovery in 'refresh' and
'tabula_rasa' modes for newly created hosts on remote sites.
This werk successfully resolves this subsequent issue.

Change-Id: I9116ab537b47dff9da4a46180a5e41bb08559be1