16371 FIX sla: range field for predefined_time_range parameter is mandatory
Before this werk, when the sla was computed for a predefined time
range without specifying the range field, an error status 500
Internal Server Error was returned. This werk solves the problem
by checking for the existence of the range field, and if it does
not exist, the endpoint returns an error status 400 - Bad Request.

Change-Id: Id77482854c7dd6d213280636ebaa3c50b476c6bf