cisco_temperature: Improve admin state detection
Previously the detection of the sensor admin state was implemented by comparing the interface description ifDescr with the sensor name entPhysicalName and using the interface admin state.
This does not work for different Cisco sytems, as e.g. Catalyst and ASR.
The mapping between interface and sensor is now attempted via the entAliasMappingTable (OID: 1.3.6.1.2.1.47.1.3.2.1).
For Cisco ASR/NX-OS a further mapping between containers and corresponding subslots is made.

In case the entAliasMappingTable is not available or could not used for mapping, the old version is used as a fallback.

SUP-16428
SUP-16903
SUP-18398

Change-Id: Id5d7fabb113fb3affde169413c9d4207b8216e0d