get_average: Use time since starting to average for short horizons
Correctly starting an exponential moving average is tricky, as
at the beginning, the last average is weighed too strongly
in respect to the information contained therein. This can lead
to initial values skewing the average for a long time until it
eventually converges.

To get to meaningful values more quickly, we replace the previously
used adjustment of the backlog weight by simply setting our backlog
horizon to min(time since starting to average, configured horizon)

Closes #675.

Co-Authored-By: Moritz Kiemer <moritz.kiemer@checkmk.com>
Change-Id: I4164202aa48df45a77579f4338edb4725a658152