14223 Introduced topology visualization
The topology visualization is a new feature that allows the visualization of complex interconnected networks.
A simple example for this visualization is the parent/child topology. The new mechanism that comes with this werk allows the linking of external data with the data of the monitoring core.
When it comes to the display, you simply define some starting points via the filter form.
Based on these, the topology visualization then builds a mesh of incoming and outgoing connections.

The type of external data might be
* Netstat, showing connections between the interfaces/ips/ports
* LLDP/CDP, showing the network neighbors

There is a common data format specification for all external data.
So you just can create your own data file which provides information about the relationships between hosts, services or generic objects which are not linked to the core.
If you drop this file into a specific folder, the visualzation will handle the rest. There is no need to write python code.

Right now you can configure
* Objects - either linked to an entity in the core or some standalone object
* Icons/emblems which should be added to the object
* Connections between objects
* Line style/color of specific connections

Since this is a quite visualization heavy topic and hard to explain only via text, feel free to check out the
[thread](https://forum.checkmk.com/t/network-visualization-now-in-version-2-3/44467) in our checkmk forum

We will also publish a blog article in the coming weeks

```
Important:
The visualization only works if external data is provided in a special folder.
At the moment these are not created by Checkmk, but come from external MKP developments.
```

Change-Id: Ia96b73a52f59537acf290aa96bca8ca7fef8fdf4