omd update: delete ca creation
We want to backup some of the data, which is modified during `omd
update`. If we can remove the update step, we don't have to backup the
data.

This was introduced update step was introduced  in commit

4638bf2545197ddafa90761be11bb06d13522215

All sites with version 2.2.0 have a site CA. Moreover, the corresponding
hook LIVESTATUS_TCP_TLS is already set to a value. This means the
initialization no longer has to be done on update.

CMK-15866

Change-Id: Ifb157f3b2bdeaf521726bd2003e48eace82e3f58