unit-test: Re-design memory overflow tests.
Memory overflow tests led to sudden assignment of ~300 MiBs in RAM.
These tests have been redesigned to simulate usage of memory in RAM.

Assignment of memory in RAM was being done to validate "memory-usage" mechanism in the code.
An aditional and independent test has been added to validate this mechanism.

CMK-14178

Change-Id: Ic28c0b8198ed50308baa2f9d4ed320ac67b2883d