16579 downtimes: Added 'modify downtimes' endpoint
With this Werk a new endpoint for modifying downtimes is incorporated.
It is possible to modify the comment and the end timestamp as in the GUI.
The selection of downtimes to modify can be done by id, by query, or by
hostname and service description.

Method: PUT
URL: domain-types/downtime/actions/modify/invoke

CMK-16144

Change-Id: I216e77a8b7e7821393f48a0cd9ec77503f20addc