omd update: remove apache update enforcement
We want to backup some of the data, which is modified during `omd
update`. If we can remove the update step, we don't have to backup the
data.

Enforcing the new apache hooks was done, when updating to the 2.2.0. See
the commit

412dc99c33b48b1816334efe929084eb1c043874

This means all apache hooks are up-to-date and don't need to upgrade
during the upgrade to 2.3.0.

CMK-15866

Change-Id: I932ef8433db6bffb68cf921c92e83c297194a676