omd update (1): rollback skeleton files
The command `omd update` must abort, if the pre-update checks fail.
Different update steps depend on each other, so reordering these steps
is not trivial. Instead we make it possible to rollback changes, which
have been made before the pre-update checks. An additional benefit is
that if the user cancels the update (using the `abort` option or CTRL-C),
then the site can be in a consistent state. Note that failures after the
pre-update checks are beyond the scope of the 2.3.0. These type of
failures are not covered by the rollback mechanism.

This change allows rollbacks of any change done by iterating over the
skeleton files.

CMK-15866

Change-Id: Ib06fef303c5faa2125d07d7cf5bedd8ac88d364b