15725 FIX Cleanup old Microcore config during update procedure
This change prevents a problem which might occur in case the `omd update` did
not finish successfully. In this situation, the Microcore might be started with
a configuration file from the previous version. This could lead to unexpected
behavior.

Instead of keeping the old configuration, the update procedure now deletes the
file which makes the Microcore fail during startup with a more helpful error
message.

CMK-16020

Change-Id: I5c7a326d9269ada9ae0f4f8b4e48610bdaffabdb