omd update (2): rollback version link
The command `omd update` needs to abort, if the pre-update checks fail.
Different update steps depend on one another, so reordering these steps
is not trivial. Instead we make it possible to rollback changes, which
have been done before the pre-update checks. An additional benefit is
that if the user aborts his update (using the `abort` option or CTRL-C),
then the site can be in a consistent state. Note that failures after the
pre-update checks are beyond the scope of the 2.3.0. These type of
failures are not covered by the rollback mechanism.

This change allows rollbacks of changes to the version link.

CMK-15866

Change-Id: Ia08f06e0895a239453906498cfed82fbcd7ec9ef