omd update (3): manage more update steps
The command `omd update` must abort, if the pre-update checks fail.
Different update steps depend on each other, so reordering these steps
is not trivial. Instead we make it possible to rollback changes, which
have been made before the pre-update checks. An additional benefit is
that if the user cancels the update (using the `abort` option or CTRL-C),
then the site can be in a consistent state. Note that failures after the
pre-update checks are beyond the scope of the 2.3.0. These type of
failures are not covered by the rollback mechanism.

The rollback mechanism is extended to three more update steps. None of
these need a rollback:
* `save_version_meta_data` is unaffected by a failed execution and does
  not concern user data.
* `site.load_config` and `load_defaults` are state free. This was
  verified by checking all hooks in 2.3.0.
* `legacy_agent_ca` is incorrectly symlinked. The new symlink is
  appropriate for 2.2.0 and 2.3.0.

CMK-15866

Change-Id: I91830caaaf7e90ee8b2ecacf080cfb3208c926ed