omd update (4): rollback calls to config hooks
The command `omd update` must abort, if the pre-update checks fail.
Different update steps depend on each other, so reordering these steps
is not trivial. Instead we make it possible to rollback changes, which
have been made before the pre-update checks. An additional benefit is
that if the user cancels the update (using the `abort` option or CTRL-C),
then the site can be in a consistent state. Note that failures after the
pre-update checks are beyond the scope of the 2.3.0. These type of
failures are not covered by the rollback mechanism.

This change allows rollbacks update actions executed by the config hooks.
These bash scripts contain update actions specific to the omd package.
This mechanism is maintained by hand in 2.3.0, but having an appropriate
solution here is needed in the future.

CMK-15866

Change-Id: Ib66b2548c5504efadb18d0514155442f508db088