integration test: Fix test_check_http
The test test_check_http was disabled for some distributions becaused it
failed so far. The reason for this was that the host was configured to
use an agent, which caused an error message if no agent was available.

This change reconfigures the test host so that no agent is used and
unskips the test for all distributions.

Change-Id: Iae5a77568f3e7ca509ddbfccad5889949f8f0308