Reset failed login attempts to 0 when unlocked
Previously, when a user was locked due to too many failed attempts, the
failed attempts counter was not reset when and admin unlocked the user.
As a result, the user had only one attempt to login before being locked
out again.
The counter is reset now so the unlocked user has the full number of
failed attempts at their disposal.

Change-Id: I8facb4054ced1ae06dd7aebc01d329a875e449cf