omd update(5): umount `tmpfs` on failure
The command `omd update` must abort, if the pre-update checks fail.
Different update steps depend on each other, so reordering these steps
is not trivial. Instead we make it possible to rollback changes, which
have been made before the pre-update checks. An additional benefit is
that if the user cancels the update (using the `abort` option or CTRL-C),
then the site can be in a consistent state. Note that failures after the
pre-update checks are beyond the scope of the 2.3.0. These type of
failures are not covered by the rollback mechanism.

`tmpfs` is unmounted if `omd update` is called. This even applies if the
customer does not commit to the update. The rollback needs to umount
`tmpfs` if it has been repopulated based on the new version.

CMK-15866

Change-Id: I3f2be64c3a277ab57a1d3c07d803a13763757868