Make parse_secret's type more honest.
It *has* more requirements than its previous signature stated, and testing
these requirements within that function is an anti-pattern.

If we really want more validation for the value part of the secret, that
should live in e.g. __post_init__()s of StoredSecret and PlainSecret, but
there is not really much point in doing so: If we deem this necessary, we
would need *tons* of other validations and should probably switch to
pydantic models directly.

Fixed a wrong type annotation in the cisco plugin on the way.

Change-Id: I550e75e66d13ec9cad47e7531b5a890b1068f44f