Removed superfluous "yield from"s.
We already have an Iterable[SpecialAgentCommand], and we need to return the
exact same type, so there is no need for a "yield from". If it already was
e.g. a generator, it stays one. And if it was e.g. a list, there is no
point in "lazyfying" it.

Change-Id: Iabca8299170b64f252c667e13db85168c9551235