16501 omd: trigger openapi-spec generation job during start, restart & reload
Werk 15724 introduces a mechanism for regenerating the API specification,
which, in certain instances, is executed as a background job due to the
potential time required for completion. This update modifies the omd start,
restart, and reload commands to initiate this specific background job upon
execution. Users should be aware that triggering the job does not result
in immediate availability of the updated documentation; there may be a
delay before the documentation is updated. This saves the user from having
to trigger and wait for the specification regeneration manually in case
relevant changes have been made outside the user interface and the apache
process needs to be restarted.

Based on werk 15724 the specification is now updated in these situations:
* post-create hook: Create the initial spec after a site has been created
* post rename action: Update the spec after a site has been copied, restored or renamed
* update-config action: Update the spec after the site has been updated
* omd apache: Update the spec when the apache process is started, restarted or reloaded

Change-Id: I769d927192e1ec03fd2f433839caf7eca7c733bb