Follow up for werk #15725
Since ff0b047743 the Microcore config is deleted at the beginning of
`cmk-update-config`. In the context of `omd update` it worked as
expected because omd cared for recreating the Microcore config. However,
when `cmk-update-config` is called directly, the Microcore config is not
recreated.

With this change we move the recreation of the Microcore config to
`cmk-update-config` so that the deletion and recreation is managed by
the same program.

CMK-16020

Change-Id: If0d6f175d47a77b3ad7e2111a38a3a0567a7ee1d