omd update(6): allow restoring after pre-update
The command `omd update` must abort, if the pre-update checks fail.
Different update steps depend on each other, so reordering these steps
is not trivial. Instead we make it possible to rollback changes, which
have been made before the pre-update checks. An additional benefit is
that if the user cancels the update (using the `abort` option or CTRL-C),
then the site can be in a consistent state. Note that failures after the
pre-update checks are beyond the scope of the 2.3.0. These type of
failures are not covered by the rollback mechanism.

We allow to rollback changes after a pre-update fails.

CMK-15866

Change-Id: I4bd984afbc7fc6f7dfa45ae01b40664da3a3c3b5