kube: fix ValidationError() after update
Since Werk #15872 piggyback data is persisted across updates. This imposes a new requirement on the special agent check plugins: Before they were only required to be compatible with the current version of the special agent sections. However, on the first check cycle after an update the old sections will be forwarded to the check plugins, since these sections persisted across updates.

In this instance, `kube_node_conditions_v1` changed its format. This was done to implement Werk #15646. By changing the section name, we ensure this old data is ignored in 2.3.0.

CMK-15963

Change-Id: I13b8eea2e98d9f1288a5991ae96ebb7b1312f219