cmk update: fix keepold,install conflictmode
The modes `keepold` and `install` are not allowed to prompt the user
interactively (see help). This problem was introduced in the change with
Change-Id:

I408f9e19f1f2666e715bdaa8c3cbf862f0aa8648

In particular, the `install` options needs to run without a tty, since
it is used by the appliance.

Change-Id: I42c0a9da3ba8168931dc66d648f5bb420ffbb478