agent_based.v1: drop TODO
The function has been part of the check API since
1.2.6. Unless there is something fundamentally
broken about it we should keep the name and the
functionality.

It is designed for non-developers, and aims to
provide a simple interface for the most common
use cases.

Unlike the early versions of this function, in
the current agent_based APIs it is a pure
convenience function that dosen't require any
magic global context.

Advanced users that don't agree with the name
or scope or have special requirements can
implement drop-in replacements that better suit
their needs.

Change-Id: I437c67098d7ca82f984630aaabb5b6e89b1630ea