16526 FIX password: the response schema now matches what is returned
This werk addresses an issue with the REST-API password endpoint
response.  The response schema listed the title and the ident as
fields that should be returned but we were not returning them as
part of the password object.  We now do.

Also, the members field was returning confusing and irrelevant
information so this has been removed.

Change-Id: I763f111fa1ab1fdd56133752a9e76bcdb9c96794