16526 FIX password: the response schema now matches what is returned
This werk addresses an issue with the REST-API password endpoint
response. The response schema listed the title and the ident as
fields that should be returned but we were not returning them as
part of the password object. These have now been removed from the
schema.

Also, the members field was returning invalid information and
hence has been removed.

SUP-19165

Change-Id: I763f111fa1ab1fdd56133752a9e76bcdb9c96794