16484 FIX Crash when creating combined graphs with empty time filter
When creating a combined graph with an empty time filter (e.g. Last service check),
the creation of the combined graph would crash.

This behavior is not consistent with the view filtering behavior,
where the filter is not applied if it is empty.

Now the filter is not applied to combined graphs either.

Change-Id: Iabe1a96755802692f0315afdc11d480f43d3aad3