15514 check_httpv2: Introduce a reworked way to test web sites
The legacy http monitoring plugin caused quite some trouble over the last
years. This included lots of effort to add features or just simply fixing
bugs.

With the new plugin, the functionality is moved to maintainable and
extendable code completely under control of Checkmk. This means also
breaking changes with the old plugin:

* Some metrics are not available anymore as it has been known. We
  discovered that these are simply not directly understandable. Instead we
  will add metrics as needed in the future. Some metrics will already be
  added in this first release
* Some functionality has been a workaround and is now implemented directly
  into the new plugin. This makes it hard to migrate rules automatically.
* Users are now able to decide on their own which functionality should be
  in an own service. This means, that it is now possible to test the
  certificate validity and response times in one service, if needed.
* User are able to configure multiple http checks within one rule. For
  that standard settings are provided used for all end points to test
  unless overwritten in the configuration of the endpoint. Migrating
  manually makes absolute sense here.

Please note that we will not remove the old plugin for now. We understand
that you need some time to migrate your configurations. Nethertheless, we
will deprecate the old plugin and eventually remove it from Checkmk.

Change-Id: Id0cece13c7af9da96108c565623fe964ffa95f10