16503 bulk discovery: remove mode parameter and introduce options
Starting with the 2.3, the bulk discovery mode has been split up
for more granular control regarding the discovery execution. This
werk now introduces the same approach to the equivalent REST API
bulk discovery endpoint introducing the 'options' field. The previous 'mode'
field is deprecated with the 2.3 and will be removed in 2.4. The user should
migrate existing scripts to make use of the 'options' field.

Change-Id: Ifff304b294f69bd33402fd1134c572eddea97cf0