Add werk command to preview html rendering
Markdown rendering is not 100% standardized and some choices seem
unintuitive. Its possible that the markdown is rendered as expected on
github, but the rendering on our homepage looks broken.

The concrete problem is list rendering, and a good discussion about this
can be found at:
https://talk.commonmark.org/t/blank-lines-before-lists-revisited/1990

We decided to not tweak the markdown rendering provided by our python
library, but to provide a tool to check the rendering.

You can now use the following command to check the rendering of
your werk:

./werk preview 16259 > /tmp/index.html && firefox /tmp/index.html

CMK-16370

Change-Id: I326f8c5a36933cd145201516d83a987a553d4078