16022 bazel_cache: Monitor Bazel Remote Cache
With this werk, it is possible to monitor a Bazel Remote Cache instance.
Processing data of the optional metrics and default status endpoint.

This feature consists of a special agent, agent_bazel_cache, and a new
check, bazel_cache. The special agent can be configured via the rule
Bazel Cache Metrics.

CMK-16546

Change-Id: I2a3ce872df9fa247c290385e423d56d19f1d31a0