16617 Use session specific key for ValueSpec encryption
When a user edits a configuration e.g. for a special agent with an explicit password the complete configuration is transfered to the user.
To not reveal the password in cleartext this field is encrypted.

The key for that encryption was previously to this Werk shared amongst all users (a salt was used though).
With this Werk every user session has now a secret dedicated to this encryption so the key is rotated often and not shared amongst other users.

CMK-11925

Change-Id: I1ec79d9ce50845c86088196b00652b77af61a356